# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pharmaceutical RAG Chat API",
        "description": "\n\uc758\uc57d\ud488 RAG \ucc57\ubd07 \ubc31\uc5d4\ub4dc API\uc785\ub2c8\ub2e4.\n\n## \uc778\uc99d\n- \ucfe0\ud0a4 \uae30\ubc18 \uc138\uc158(`session` \ucfe0\ud0a4, HTTPOnly).\n- `POST /auth/register` \ub610\ub294 `POST /auth/login` \uc131\uacf5 \uc2dc \uc790\ub3d9\uc73c\ub85c \ucfe0\ud0a4\uac00 \uc124\uc815\ub429\ub2c8\ub2e4.\n- \ubaa8\ub4e0 \ubcf4\ud638\ub41c \uc5d4\ub4dc\ud3ec\uc778\ud2b8\ub294 \uc774 \ucfe0\ud0a4\ub97c \uc790\ub3d9 \ud3ec\ud568\ud569\ub2c8\ub2e4(`credentials: \"include\"`).\n\n## \uad8c\ud55c\n- `user` \u2014 \uc77c\ubc18 \uc0ac\uc6a9\uc790: \ucc44\ud305, \ubcf8\uc778 \ud504\ub85c\ud544 \uc870\ud68c/\uc218\uc815.\n- `admin` \u2014 \uad00\ub9ac\uc790: \uc704 \ubaa8\ub4e0 \uad8c\ud55c + \ud504\ub86c\ud504\ud2b8 \uc0dd\uc131/\ubc84\uc804 \uad00\ub9ac/\ud65c\uc131\ud654.\n\n## \ucc44\ud305 \uc2a4\ud2b8\ub9ac\ubc0d\n- `POST /chat/sessions/{session_id}/messages`\ub294 `text/event-stream` (SSE)\uc73c\ub85c \uc751\ub2f5\ud569\ub2c8\ub2e4.\n- \uc774\ubca4\ud2b8 \uc885\ub958: `token`, `tool`, `done`, `error`.\n",
        "version": "1.0.0"
    },
    "paths": {
...
        "/chat/sessions": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "\ub0b4 \ub300\ud654 \uc138\uc158 \ubaa9\ub85d",
                "description": "\ud604\uc7ac \ub85c\uadf8\uc778\ub41c \uc0ac\uc6a9\uc790\uac00 \uc18c\uc720\ud55c \ub300\ud654 \uc138\uc158\uc744 \ucd5c\uc2e0 \uac31\uc2e0\uc21c\uc73c\ub85c \ubc18\ud658\ud569\ub2c8\ub2e4.",
                "operationId": "list_sessions_chat_sessions_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/SessionPublic"
                                    },
                                    "type": "array",
                                    "title": "Response List Sessions Chat Sessions Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "\ub300\ud654 \uc138\uc158 \uc0dd\uc131",
                "description": "\uc0c8 \ub300\ud654 \uc138\uc158\uc744 \ub9cc\ub4ed\ub2c8\ub2e4. `title`\uc744 \uc0dd\ub7b5\ud558\uba74 \uae30\ubcf8 \uc81c\ubaa9\uc774 \ubd80\uc5ec\ub429\ub2c8\ub2e4.",
                "operationId": "create_session_chat_sessions_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSessionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/chat/sessions/{session_id}/messages": {
            "get": {
                "tags": [
                    "chat"
                ],
                "summary": "\uc138\uc158 \uba54\uc2dc\uc9c0 \uc870\ud68c",
                "description": "\ud574\ub2f9 \uc138\uc158\uc5d0 \uc800\uc7a5\ub41c \uba54\uc2dc\uc9c0\ub97c \uc2dc\uac04\uc21c\uc73c\ub85c \ubc18\ud658\ud569\ub2c8\ub2e4. \ubcf8\uc778 \uc18c\uc720 \uc138\uc158\ub9cc \uc870\ud68c \uac00\ub2a5\ud569\ub2c8\ub2e4.",
                "operationId": "list_session_messages_chat_sessions__session_id__messages_get",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MessagePublic"
                                    },
                                    "title": "Response List Session Messages Chat Sessions  Session Id  Messages Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "chat"
                ],
                "summary": "\uba54\uc2dc\uc9c0 \uc804\uc1a1 (SSE \uc2a4\ud2b8\ub9ac\ubc0d \uc751\ub2f5)",
                "description": "\uc0ac\uc6a9\uc790 \uba54\uc2dc\uc9c0\ub97c \uc804\uc1a1\ud558\uace0, \uc5b4\uc2dc\uc2a4\ud134\ud2b8\uc758 \uc751\ub2f5\uc744 **Server-Sent Events** \ub85c \uc2a4\ud2b8\ub9ac\ubc0d\ud569\ub2c8\ub2e4.\n\n\ud504\ub860\ud2b8\uc5d4\ub4dc\ub294 `fetch` + ReadableStream \ub610\ub294 `EventSource`\ub85c \uc218\uc2e0\ud558\uc138\uc694. \ubcf8 \uc751\ub2f5\uc740 JSON\uc774 \uc544\ub2d9\ub2c8\ub2e4.",
                "operationId": "post_message_chat_sessions__session_id__messages_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server-Sent Events \uc2a4\ud2b8\ub9bc. \uac01 \uc774\ubca4\ud2b8\ub294 `event:` \uc904\uacfc JSON `data:` \uc904\ub85c \uad6c\uc131\ub429\ub2c8\ub2e4.\n\n**\uc774\ubca4\ud2b8 \ud0c0\uc785:**\n- `token` \u2014 `{\"text\": \"...\"}` \uc5b4\uc2dc\uc2a4\ud134\ud2b8\uac00 \uc0dd\uc131\ud55c \ud1a0\ud070 \uc870\uac01\n- `tool` \u2014 `{\"phase\": \"start|end\", \"tool_name\": \"...\", \"input\": {...}}` \ub3c4\uad6c \ud638\ucd9c \uc9c4\ud589 \uc0c1\ud669\n- `done` \u2014 \uc2a4\ud2b8\ub9bc \uc815\uc0c1 \uc885\ub8cc. `{\"message_id\": int}`\n- `error` \u2014 `{\"message\": \"...\"}` \uc5d0\ub7ec \ubc1c\uc0dd",
                        "content": {
                            "text/event-stream": {
                                "example": "event: token\ndata: {\"text\": \"\uc548\ub155\"}\n\nevent: token\ndata: {\"text\": \"\ud558\uc138\uc694\"}\n\nevent: done\ndata: {\"message_id\": 42}\n\n"
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateSessionRequest": {
                "properties": {
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    }
                },
                "type": "object",
                "title": "CreateSessionRequest"
            },
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail",
                        "description": "\uc0ac\ub78c\uc774 \uc77d\uc744 \uc218 \uc788\ub294 \uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "examples": [
                            "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse",
                "description": "\ubaa8\ub4e0 4xx/5xx \uc751\ub2f5\uc5d0\uc11c \uc0ac\uc6a9\ud558\ub294 \uacf5\ud1b5 \uc5d0\ub7ec \uc2a4\ud0a4\ub9c8."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MessageBody": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "MessageBody"
            },
            "MessagePublic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "tool_calls": {
                        "anyOf": [
                            {
                                "items": {},
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tool Calls"
                    },
                    "prompt_version_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Prompt Version Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "role",
                    "content",
                    "created_at"
                ],
                "title": "MessagePublic"
            },
...
            "SessionPublic": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "created_at",
                    "updated_at"
                ],
                "title": "SessionPublic"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    },
...
}